requirements.txt
numpy>=1.21.0,<2.0
opencv-python>=4.5.0
matplotlib>=3.5.0
uvicorn[standard]
opencv-python
pydantic>=2
fastapi
Pillow
requests
torch
transformers
torchvision
accelerate 
bitsandbytes
httpx
ultralytics
scipy